Validation, unit tests
Added validation and tests for add requests

+++ src/main-api/Domain/Models/Ingredient.cs
@@ -6,7 +6,7 @@ public class Ingredient
     [Required]
     public required Product Product { get; set; }
     [Required]
-    public required double Weight { get; set; }
+    public required decimal Weight { get; set; }
     [Required]
     public required Unit Unit { get; set; }
 }
